refactor(engine): extract SongGenerationProcessor + quality unit tests (Phase 2 PR-8)
Verbatim move of nine GameEngine song-generation methods into
shared/engine/processors/SongGenerationProcessor.ts (processRecordingProjects,
shouldGenerateProjectSongs, generateWeeklyProjectSongs, getSongsPerWeek,
generateSong, calculateEnhancedSongQuality, generateSongMood,
generateSongEconomicInsight, generateProjectCompletionSummary). Bodies moved
wholesale with only this. rebound to the injected WeekContext; RNG draw order in
the quality formula (outlier roll + normal variance) preserved exactly.

Engine keeps two same-signature delegates: processRecordingProjects (called from
advanceWeek) and calculateEnhancedSongQuality (still reached by the engine
diagnostics path and scripts/verification/test-song-quality-scenarios.ts). The
other seven had no external callers and are removed. calculateBudgetQualityBonus
and calculateSongCountQualityImpact are unrelated engine->FinancialSystem
passthroughs and were left untouched.

Adds tests/engine/song-quality.test.ts (17 tests) exercising the crown-jewel
quality formula directly against the processor with fixed/scripted seeds:
determinism, talent/producer-tier/time-investment/mood/popularity/song-count
factor boundaries, floor/ceiling clamps (25/98), and the breakout / critical-
failure / normal-variance outlier branches. Closes review-gap W3 for quality.

Golden master 8/8 byte-identical (zero snapshot updates); full suite 670/670.
